+++ src/Api/Controllers/AssistantController.cs
@@ -12,6 +12,8 @@ using TheKrystalShip.Api.Services.Leaves;
 using TheKrystalShip.KGSM.Auth;
 using TheKrystalShip.KGSM.Auth.Discord;
 
+using TheKrystalShip.KGSM.Auth.Sessions;
+
 namespace TheKrystalShip.Api.Controllers;
 
 /// <summary>